+ added overlapping preset selection
More than one preset can now be selected at the same time, which
will result into a union of the Home panel options of all the
selected presets.
For example, if the YouTube's "Sneaky" preset is selected first,
and then the "Cozy" preset is selected, the result will be the
selection of both the presets' Home panel options.
In this situation, if a preset is deselected, only its Home panel
options will be deselected.
If, after one or more presets have been selected, any Home panel
option is deselected, all selected presets will be deselected.

v2.5.5