+ added option to remove extra HTML tags for YouTube
One option (the one that deleted only <iframe> elements) has been
modified to remove all the tags in <head> that are not essential
for the appearance of the webpage.
A new option has been added to remove all the tags inside <body>
(both immidiate and nested children) that are not essential for a
distraction-free appearance of the webpage, such as <script>,
<meta>, <link>, and YouTube's custom tags.
It should be noted that at least the <script> tags should only be
removed after loading the video. Otherwise, the video may not load.

v2.6.7