+ modified YouTube option target
If the merchandise section exists after the video details on YouTube,
the "Video details" options would only remove the details, which would
push the merchandise section up, sometimes inconveniently covering the
video partially.

To fix this, the merchandise section has been included into the target
of the "Video details" option, so that it will now also delete the
section.

v2.7.4